Moved service implementations to different package
